Make InPlaceABNSync compatible with future Pytorch versions
Summary:
Future Pytorch versions will change the way `torch.distributed.group.WORLD` is initialized. This will break the current implementation of `InPlaceABNSync`, such that it would revert to un-synchronized `InPlaceABN` behavior unless `set_group()` is called. To prevent this, the constructor of `InPlaceABNSync`, as well as the `inplace_abn_sync` function are changed to accept `None` as a default `group` parameter, which will now trigger synchronization across the default distributed group. This change should preserve backward compatibility with current and previous versions of Pytorch.

Additionally, the way batch statistics are synchronized across distributed workers is now changed to be closer to Pytorch's `SyncBatchNorm`.

Reviewed By: rotabulo

Differential Revision: D25899607

fbshipit-source-id: 95cdab20fcdd0cc0f47bca0af16bc2ff7ac01dc9